now logs long jumps

+++ include/libRR_c.h
@@ -18,7 +18,7 @@ void libRR_log_instruction_z80(uint32_t current_pc, const char* c_name, uint32_t
 void libRR_log_instruction_z80_s_d(uint32_t current_pc, const char* c_name, uint32_t instruction_bytes, int number_of_bytes, const char* source, const char* destination);
 const char* libRR_log_jump_label(int32_t offset);
 void libRR_log_memory_read(int8_t bank, int32_t offset, const char* type, uint8_t byte_size, char* bytes);
-void libRR_log_rom_read(int8_t bank, int32_t offset, const char* type, uint8_t byte_size, char* bytes);
+void libRR_log_rom_read(int16_t bank, int32_t offset, const char* type, uint8_t byte_size, char* bytes);
 void libRR_gameboy_log_memory_read(int32_t offset, const char* type, uint8_t byte_size, char* bytes);
